Inline glob-to-code-search mapping in githubvfs
Remove the GlobQueryResolver interface and LLM warm-up path.
Known discovery patterns map directly to GitHub code search
queries inside githubvfs.
