--- pkg/discovery/github/prompts/glob_code_search.txt.tmpl
@@ -1,33 +1,0 @@
-You translate compliance discovery vfs glob patterns into GitHub code search queries.
-
-Each glob pattern uses `*` as the repository name segment, for example:
-- `*/SECURITY.md` means "SECURITY.md at the root of any repository"
-- `*/.github/workflows/*.yml` means workflow YAML files under `.github/workflows`
-
-For each pattern below, return a GitHub code search query fragment WITHOUT the org qualifier.
-Use GitHub search qualifiers such as path:, filename:, and extension:.
-Do not include repo: or org: in search_fragment.
-
-Return JSON only with this shape:
-{
-  "queries": [
-    {
-      "pattern": "*/SECURITY.md",
-      "search_fragment": "filename:SECURITY.md"
-    }
-  ]
-}
-
-Rules:
-- cover every pattern from the input list exactly once
-- prefer precise qualifiers over broad keyword search
-- for lockfiles and dotfiles, use filename: when possible
-- for nested paths, combine path: and filename: as needed
-- when a static hint exists, you may reuse or improve it
-- skip patterns that cannot be expressed in GitHub code search
-
-Static hints (optional reference, not mandatory):
-{{.StaticHintsJSON}}
-
-Patterns to resolve:
-{{.PatternsJSON}}
