+++ pkg/discovery/github/scanner.go
@@ -31,7 +31,6 @@ type discoveryScanner struct {
 	logger              *log.Logger
 	scopes              scopeSet
 	repoClassifier      RepoClassifier
-	globResolver        GlobQueryResolver
 	repoClassifications map[string]RepoClassification
 }
 
@@ -41,7 +40,6 @@ func newDiscoveryScanner(
 	conn connector.Connection,
 	logger *log.Logger,
 	repoClassifier RepoClassifier,
-	globResolver GlobQueryResolver,
 ) *discoveryScanner {
 	api := newAPIClient(httpClient)
 
@@ -49,19 +47,14 @@ func newDiscoveryScanner(
 		repoClassifier = DefaultRepoClassifier()
 	}
 
-	if globResolver == nil {
-		globResolver = DefaultGlobQueryResolver()
-	}
-
 	return &discoveryScanner{
 		api:            api,
-		fs:             newGitHubFS(api, org, globResolver),
+		fs:             newGitHubFS(api, org),
 		org:            org,
 		conn:           conn,
 		logger:         logger,
 		scopes:         newScopeSet(conn),
 		repoClassifier: repoClassifier,
-		globResolver:   globResolver,
 	}
 }
